refactor: BlackJackGame을 컨트롤러로 취급 후 뷰 의존, 뷰 메서드가 구현한 인터페이스 삭제
