--- src/main/java/blackjack/domain/game/BettingInjector.java
@@ -1,6 +1,0 @@
-package blackjack.domain.game;
-
-@FunctionalInterface
-public interface BettingInjector {
-	int inject(String name);
-}
