--- src/main/java/blackjack/domain/game/CardsChecker.java
@@ -1,10 +1,0 @@
-package blackjack.domain.game;
-
-import java.util.List;
-
-import blackjack.domain.card.Card;
-
-@FunctionalInterface
-public interface CardsChecker {
-	void check(String name, List<Card> cards);
-}
