--- src/main/java/blackjack/domain/game/HitRequester.java
@@ -1,6 +1,0 @@
-package blackjack.domain.game;
-
-@FunctionalInterface
-public interface HitRequester {
-	boolean request(String name);
-}
